# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.5.0",
    "title": "FCTrading API client example",
    "description": "Example for test eassssy"
  },
  "host": "localhost:3011",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/verifyCode": {
      "get": {
        "tags": [
          "AUTH"
        ],
        "description": "",
        "parameters": [
          {
            "name": "twoFaType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/newOrder": {
      "get": {
        "tags": [
          "ORDER"
        ],
        "description": "",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "type": "string"
          },
          {
            "name": "requestID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "instrumentID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "market",
            "in": "query",
            "type": "string"
          },
          {
            "name": "buySell",
            "in": "query",
            "type": "string"
          },
          {
            "name": "orderType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "type": "string"
          },
          {
            "name": "quantity",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stopOrder",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stopPrice",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stopType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stopStep",
            "in": "query",
            "type": "string"
          },
          {
            "name": "lossStep",
            "in": "query",
            "type": "string"
          },
          {
            "name": "profitStep",
            "in": "query",
            "type": "string"
          },
          {
            "name": "channelID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userAgent",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/derNewOrder": {
      "get": {
        "tags": [
          "ORDER"
        ],
        "description": "",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "type": "string"
          },
          {
            "name": "requestID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "instrumentID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "market",
            "in": "query",
            "type": "string"
          },
          {
            "name": "buySell",
            "in": "query",
            "type": "string"
          },
          {
            "name": "orderType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "type": "string"
          },
          {
            "name": "quantity",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stopOrder",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stopPrice",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stopType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "stopStep",
            "in": "query",
            "type": "string"
          },
          {
            "name": "lossStep",
            "in": "query",
            "type": "string"
          },
          {
            "name": "profitStep",
            "in": "query",
            "type": "string"
          },
          {
            "name": "channelID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userAgent",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/modifyOrder": {
      "get": {
        "tags": [
          "ORDER"
        ],
        "description": "",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "type": "string"
          },
          {
            "name": "orderid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "market",
            "in": "query",
            "type": "string"
          },
          {
            "name": "instrumentid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "type": "string"
          },
          {
            "name": "quantity",
            "in": "query",
            "type": "string"
          },
          {
            "name": "buysell",
            "in": "query",
            "type": "string"
          },
          {
            "name": "orderType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userAgent",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/derModifyOrder": {
      "get": {
        "tags": [
          "ORDER"
        ],
        "description": "",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "type": "string"
          },
          {
            "name": "orderid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "market",
            "in": "query",
            "type": "string"
          },
          {
            "name": "instrumentid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "price",
            "in": "query",
            "type": "string"
          },
          {
            "name": "quantity",
            "in": "query",
            "type": "string"
          },
          {
            "name": "buysell",
            "in": "query",
            "type": "string"
          },
          {
            "name": "orderType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userAgent",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/cancelOrder": {
      "get": {
        "tags": [
          "ORDER"
        ],
        "description": "",
        "parameters": [
          {
            "name": "orderid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "account",
            "in": "query",
            "type": "string"
          },
          {
            "name": "instrumentid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "market",
            "in": "query",
            "type": "string"
          },
          {
            "name": "buysell",
            "in": "query",
            "type": "string"
          },
          {
            "name": "requestID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userAgent",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/derCancelOrder": {
      "get": {
        "tags": [
          "ORDER"
        ],
        "description": "",
        "parameters": [
          {
            "name": "orderid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "account",
            "in": "query",
            "type": "string"
          },
          {
            "name": "instrumentid",
            "in": "query",
            "type": "string"
          },
          {
            "name": "market",
            "in": "query",
            "type": "string"
          },
          {
            "name": "buysell",
            "in": "query",
            "type": "string"
          },
          {
            "name": "requestID",
            "in": "query",
            "type": "string"
          },
          {
            "name": "code",
            "in": "query",
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "userAgent",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/orderHistory": {
      "get": {
        "tags": [
          "QUERY"
        ],
        "description": "",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/orderBook": {
      "get": {
        "tags": [
          "QUERY"
        ],
        "description": "",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}